Bugfix: Vite build - remove reduntant css links in blade compontents + update tinymce config
