+++ resources/js/admin.js
@@ -1,4 +1,3 @@
-// import "../css/admin.css"
 import Sortable from 'sortablejs';
 
 // const projects = document.querySelectorAll("#project-item");
