+++ resources/views/components/includes/tinymce-config.blade.php
@@ -1,41 +1,41 @@
 <script src="{{ asset('js/tinymce/tinymce.min.js') }}" referrerpolicy="origin"></script>
 <script>
-    tinymce.init({
-        selector: 'textarea.description', // Replace this CSS selector to match the placeholder element for TinyMCE
-        plugins: 'code table lists link image',
-        toolbar: 'undo redo | link | image | blocks | bold italic | alignleft aligncenter alignright | indent outdent | bullist numlist | code | table',
-        link_default_target: '_blank',
-        file_picker_types: 'image',
-        /* and here's our custom image picker*/
-        file_picker_callback: (cb, value, meta) => {
-            const input = document.createElement('input');
-            input.setAttribute('type', 'file');
-            input.setAttribute('accept', 'image/*');
+    // tinymce.init({
+    //     selector: 'textarea.description', // Replace this CSS selector to match the placeholder element for TinyMCE
+    //     plugins: 'code table lists link image',
+    //     toolbar: 'undo redo | link | image | blocks | bold italic | alignleft aligncenter alignright | indent outdent | bullist numlist | code | table',
+    //     link_default_target: '_blank',
+    //     file_picker_types: 'image',
+    //     /* and here's our custom image picker*/
+    //     file_picker_callback: (cb, value, meta) => {
+    //         const input = document.createElement('input');
+    //         input.setAttribute('type', 'file');
+    //         input.setAttribute('accept', 'image/*');
 
-            input.addEventListener('change', (e) => {
-                const file = e.target.files[0];
+    //         input.addEventListener('change', (e) => {
+    //             const file = e.target.files[0];
 
-                const reader = new FileReader();
-                reader.addEventListener('load', () => {
-                    /*
-                      Note: Now we need to register the blob in TinyMCEs image blob
-                      registry. In the next release this part hopefully won't be
-                      necessary, as we are looking to handle it internally.
-                    */
-                    const id = 'blobid' + (new Date()).getTime();
-                    const blobCache =  tinymce.activeEditor.editorUpload.blobCache;
-                    const base64 = reader.result.split(',')[1];
-                    const blobInfo = blobCache.create(id, file, base64);
-                    blobCache.add(blobInfo);
+    //             const reader = new FileReader();
+    //             reader.addEventListener('load', () => {
+    //                 /*
+    //                   Note: Now we need to register the blob in TinyMCEs image blob
+    //                   registry. In the next release this part hopefully won't be
+    //                   necessary, as we are looking to handle it internally.
+    //                 */
+    //                 const id = 'blobid' + (new Date()).getTime();
+    //                 const blobCache =  tinymce.activeEditor.editorUpload.blobCache;
+    //                 const base64 = reader.result.split(',')[1];
+    //                 const blobInfo = blobCache.create(id, file, base64);
+    //                 blobCache.add(blobInfo);
 
-                    /* call the callback and populate the Title field with the file name */
-                    cb(blobInfo.blobUri(), { title: file.name });
-                });
-                reader.readAsDataURL(file);
-            });
+    //                 /* call the callback and populate the Title field with the file name */
+    //                 cb(blobInfo.blobUri(), { title: file.name });
+    //             });
+    //             reader.readAsDataURL(file);
+    //         });
 
-            input.click();
-        },
+    //         input.click();
+    //     },
 
-    });
+    // });
 </script>
